test: isolate Playwright visual tests (rename + placeholder) to avoid Vitest import issues
